feat: implement user recent location for search nearby recommendation place, fix: single state management on location

+++ app/src/main/java/gli/project/tripmate/data/remote/model/DetailPlaceResponse.kt
@@ -280,7 +280,7 @@ data class DetailRaw(
 	val buildingMaterial: String? = null,
 
 	@field:SerializedName("osm_id")
-	val osmId: Int? = null,
+	val osmId: Long? = null,
 
 	@field:SerializedName("name:vi")
 	val nameVi: String? = null,
